test: fix tests

+++ src/MiniJwt.Core/Models/MiniJwtOptions.cs
@@ -5,5 +5,5 @@ public class MiniJwtOptions
     public string SecretKey { get; set; } = string.Empty;
     public string Issuer { get; set; } = string.Empty;
     public string Audience { get; set; } = string.Empty;
-    public int ExpirationMinutes { get; set; } = 60;
+    public double ExpirationMinutes { get; set; } = 60;
 }
